refactor: use context for font size and ruler variables. load value to slider on text size fix
