+++ src/EulexiaFab/RulerSetup/RulerSetup.jsx
@@ -1,48 +1,14 @@
 import React, { useContext } from 'react'
-import styled from 'styled-components'
-import { RulerContext } from '../ruler-context.jsx'
 
-const RulerTop = styled.div.attrs((props) => ({
-  style: {
-    height: props.rulerPosition - props.rulerSize / 2
-  }
-}))`
-  background-color: rgba(0, 0, 0, 0.7);
-  width: 100%;
-  top: 0;
-  left: 0;
-  position: fixed;
-  pointer-events: none;
-`
+import { EulexiaContext } from '../../Context/eulexia-context.jsx'
 
-const RulerBot = styled.div.attrs((props) => ({
-  style: {
-    top: props.rulerPosition + props.rulerSize / 2
-  }
-}))`
-  background-color: rgba(0, 0, 0, 0.7);
-  width: 100%;
-  bottom: 0;
-  left: 0;
-  position: fixed;
-  pointer-events: none;
-`
-
-const RulerLine = styled.div.attrs((props) => ({
-  style: {
-    top: props.rulerPosition - props.rulerSize / 2,
-    height: props.rulerSize
-  }
-}))`
-  background-color: rgba(0, 0, 0, 0.4);
-  width: 100%;
-  position: fixed;
-  pointer-events: none;
-`
+import RulerBot from './RulerBot/RulerBot.jsx'
+import RulerTop from './RulerTop/RulerTop.jsx'
+import RulerLine from './RulerLine/RulerLine.jsx'
 
 const RulerSetup = () => {
   const { rulerEnabled, rulerSize, rulerPosition, rulerInverted } = useContext(
-    RulerContext
+    EulexiaContext
   )
 
   return (
