+++ src/EulexiaFab/index.jsx
@@ -1,6 +1,8 @@
 import React from 'react'
-import Fab from '../Fab/Fab.jsx'
 
+import { EulexiaContextProvider } from '../Context/eulexia-context.jsx'
+
+import Fab from '../Fab/Fab.jsx'
 import FontSizeAction from './FontSizeAction/FontSizeAction.jsx'
 import FontFamilyAction from './FontFamilyAction/FontFamilyAction.jsx'
 import ColorChangeAction from './ColorChangeAction/ColorChangeAction.jsx'
@@ -8,11 +10,9 @@ import TextToSpeechAction from './TextToSpeechAction/TextToSpeechAction.jsx'
 import RulerAction from './RulerAction/RulerAction.jsx'
 import RulerSetup from './RulerSetup/RulerSetup.jsx'
 
-import { RulerContextProvider } from './ruler-context.jsx'
-
 const EulexiaFab = () => {
   return (
-    <RulerContextProvider>
+    <EulexiaContextProvider>
       <RulerSetup />
       <Fab>
         <FontSizeAction />
@@ -21,7 +21,7 @@ const EulexiaFab = () => {
         <RulerAction />
         <TextToSpeechAction />
       </Fab>
-    </RulerContextProvider>
+    </EulexiaContextProvider>
   )
 }
